+++ Fait.DB/dbo/Tables/Degree.sql
@@ -1,6 +1,5 @@
-﻿CREATE TABLE [dbo].[Degree]
-(
-	[Id]	INT	IDENTITY(1, 1)	NOT NULL, 
+﻿CREATE TABLE [dbo].[Degree] (
+	[Id]	TINYINT	IDENTITY(1, 1)	NOT NULL, 
     [Name]	NCHAR(10)	NOT NULL,
 	PRIMARY KEY ([Id])
 )
